Move clouds.yaml configuration inside nodepool
Rely on nodepool module to manage clouds.yaml configuration.
This requires a simple array to be passed, such as:

clouds => [
  {
    'name': 'cloud1',
    'provider': 'hp',
    'regions': 'reg1, reg2',
    'username': 'username1',
    'pass': 'pass1',
    'auth_url': 'http://auth_url1',
  },
  {
    'name': 'cloud2',
    'provider': 'hp',
    'regions': 'reg2, reg3',
    'username': 'username2',
    'pass': 'pass2',
    'auth_url': 'http://auth_url2',
  },
]

That kind of logic is better handled by the nodepool
module, and the fact that it needs some secrets to be
populated for each cloud, makes it more difficult to move that
into an independent template.

Change-Id: I9d5853ddb03a314eeff211b4f1381e4ad7326437