Adds pip dependencies
This change adds dependencies required to build a nodepool image. The use
of the Shade module in the current release requires other modules to be
installed. This change enusures that the wrapt, dogpile.core, jsonpointer,
cliff, futures, shade, and cryptography python packages are all present
before creating a nodepool image.

Change-Id: Ibb5ee6c5da4a954f9940417342dc03095c9a715f