Remove mysql::python class inclusion
Nodepool depends on PyMySql and installs it from pip as part of its
requirements. Therefore it is no longer necessary to include the
mysql::python class, which installs the python-mysqldb package from
Ubuntu packages. Removing this class inclusion will make transitioning
to an up-to-date version of the mysql puppet module easier, as the
current version moves the mysql::python class to
mysql::bindings::python.

Change-Id: I6ac8d7433d4a0a2c591f718e3b4579b27de67e24