Allow a builder-worker to run on the nodepool host
This is behind a feature flag.

Nodepool is splitting out builder-workers so that node builds can run in
parallel and the nodepool use won't need sudo. This patch runs a single
worker on the existing nodepool.o.o node. This is the first step, then
later we will run multiple nodes runing workers.

Change-Id: I295e14d7494d859108a0e247c8106a9b673c629d