Split nodepool and zuul key
This tries to make the distinction between the nodepool user and the
zuul user clearer by putting the zuul user's public-key into the run
directory.

As explained in the comments added to the parameter; there are really
two users -- the "nodepool" private key is the one used to log into a
newly booted image and run the preparation scripts.  This key is
populated via configdrive meta-data to the new host (that's why we
have our funny little loop where we login as "ubuntu", "centos",
"cloud-user" etc until something works).

The zuul-worker public key is logically separate.  That's what zuul
will use to log into the worker node when it is attached.

i.e. in the current situation the private key ~nodepool/.ssh/id_rsa is
for the nodepool user, and the ~nodepool/.ssh/id_rsa.pub is for a
different user (zuul-worker), which is rather confusing.

This puts the zuul-worker key into /var/run/nodepool as it is sort of
ephemeral data that might change.  With
Ic9c9e415c158ad1f057b8d2aa2776dbe2bbd1e47 we can redirect the
zuul-worker DIB element to pick up this file.

Change-Id: I2a9189f18f0b09ea7b2dfa4c4bb1873a6c538e1c