Serve qcow2 images from builder http servers
Let people download our qcow2 images to aid in local reproduction of
issues. We only serve the qcow2 files as they are smallest.

Note that this should probably be updated to use a negative match
against qcow2 so that we can ensure only qcow2 is served regardless of
what image formats we are using. Unfortuantely I can't get that to work
with autoindex. If you have ideas feel free to share them.

Change-Id: Ia23e68e87d2126eac2e22d9c3fed8406e2f447cb