Fix name of libxslt dev package
`libxslt-dev` is the virtual package. Puppet can install it fine, but
can't detect that it's already installed and will attempt to reinstall
it on every run, breaking idempotency tests. Use the real name of the
package, `libxslt1-dev`, to avoid this. This name is the same for
Trusty, Xenial, and Bionic.

Change-Id: If65865c5d895559f48db325b65f3e72445ef9d4b