Don't install pip3 from distro packages
When python3-pip is installed from distro packages, it pulls in
python3-six and other python dist packages. We don't need the
distro-packaged pip since we use the latest one obtained with
get-pip.py, but this pip can't install nodepool because of the
conflicting six package from the distro. This patch exchanges the
pip::python3 class for just an ensure_packages() call which will just
make sure the python3 dev package is installed and not bother with pip.

Change-Id: I06512250829bee04e23e7652364074e023f4b2e3