Add openstacksdk logging to nodepool builder
Nodepool now uses openstacksdk in place of shade. Add a logger
for that.

Also, this makes the builder logging config look like the one in
system-config which is, accidentally, the actual logging config
being used right now. I'm going to put up another change to use
the logging config in this repo, which is the correct thing to do
for v3 of nodepool.

Change-Id: I5e6ab0a34e9e5df0ee152b33b01e5266691e6fc6